# services/orchestrator-api/openapi.yaml
openapi: 3.0.3
info:
  title: Orchestrator API
  version: "1.0.0"
  description: >
    API para ingestão de eventos OCPP, comandos, consulta de sessões, métricas,
    tarifas e faturamento.
servers:
  - url: http://localhost:3000
    description: Local
...
security:
  - ApiKeyAuth: []   # por padrão todas as rotas exigem X-API-Key
paths:
...
  /v1/events:
    get:
      tags: [Events]
      summary: Lista eventos
      parameters:
        - in: query
          name: event_type
          schema: { type: string }
        - in: query
          name: charge_box_id
          schema: { type: string }
        - in: query
          name: connector_pk
          schema: { type: integer, format: int32 }
        - in: query
          name: transaction_pk
          schema: { type: integer, format: int32 }
        - in: query
          name: id_tag
          schema: { type: string }
        - in: query
          name: from
          description: ISO 8601
          schema: { type: string, format: date-time }
        - in: query
          name: to
          description: ISO 8601
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventList'
  /v1/events/{id}:
    get:
      tags: [Events]
      summary: Busca evento por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Evento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Não encontrado

  # -------------------- OCPP ingest --------------------
  /v1/ocpp/events:
    post:
      tags: [OCPP]
      summary: Ingestão de eventos OCPP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OcppIngest'
            examples:
              start:
                value:
                  type: StartTransaction
                  transactionId: 9001
                  chargeBoxId: CB-01
                  idTag: ID-TEST
                  timestamp: "2025-08-21T12:00:00Z"
                  payload: { meterStart: 0 }
              stop:
                value:
                  type: StopTransaction
                  transactionId: 9001
                  reason: Remote
                  timestamp: "2025-08-21T12:15:00Z"
                  payload: { meterStop: 1200 }
      responses:
        '202':
          description: Aceito (novo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'
        '200':
          description: Aceito (duplicado idempotente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ack'

  # -------------------- Commands --------------------
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Ack:
      type: object
      properties:
        accepted: { type: boolean }
        idempotentDuplicate: { type: boolean }
    OcppIngest:
      type: object
      properties:
        type: { type: string }
        transactionId: { type: integer, format: int32, nullable: true }
        chargeBoxId: { type: string, nullable: true }
        idTag: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        timestamp: { type: string, format: date-time, nullable: true }
        payload:
          type: object
          additionalProperties: true
      required: [type, payload]
    Event:
      type: object
      properties:
        id: { type: string }
        created_at: { type: string, format: date-time }
        source: { type: string }
        event_type: { type: string }
        charge_box_id: { type: string, nullable: true }
        connector_pk: { type: integer, format: int32, nullable: true }
        transaction_pk: { type: integer, format: int32, nullable: true }
        id_tag: { type: string, nullable: true }
        payload:
          type: object
          additionalProperties: true
    EventList:
      type: object
      properties:
        count: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/Event'
...